feat: mapping functions
